refactor(gateway): keep one copy of the BillableUsage-to-TokenUsage map
The map is protocol-agnostic and had been copied verbatim into four
per-protocol usage modules. It belongs once, beside the TokenUsage
constructor it delegates to. Each protocol's own usage module goes back to
being `usage.ts`, and the Messages reader rejoins the extractor it had
been split from for no reason.
