+++ packages/gateway/src/data-plane/chat/chat-completions/attempt.ts
@@ -1,6 +1,6 @@
 import { chatCompletionsInterceptors } from './interceptors/index.ts';
 import type { ChatCompletionsInvocation } from './interceptors/types.ts';
-import { billableUsageFromChatCompletionsEvent } from './usage-billable.ts';
+import { billableUsageFromChatCompletionsEvent } from './usage.ts';
 import { buildUpstreamCallOptions } from '../../shared/upstream-call-options.ts';
 import { messagesAttempt } from '../messages/attempt.ts';
 import { responsesAttempt } from '../responses/attempt.ts';
